% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-item-cleaning.R
\name{clean_address_city}
\alias{clean_address_city}
\title{Clean city names}
\usage{
clean_address_city(city, keep_unknown = FALSE)
}
\arguments{
\item{city}{A character vector of city names.}

\item{keep_unknown}{Replace values for "unknown" with \code{NA}?}
}
\value{
A character vector with leading and trailing whitespace removed.
  If \code{keep_unknown} is \code{FALSE}, blanks and "UNKNOWN" are replaced
  with \code{NA}.
}
\description{
Clean city names
}
