% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-item-cleaning.R
\name{clean_address_number_and_street}
\alias{clean_address_number_and_street}
\title{Clean house number and street values}
\usage{
clean_address_number_and_street(location, keep_unknown = FALSE)
}
\arguments{
\item{location}{A character vector of house numbers and street names.}

\item{keep_unknown}{Replace values for "unknown" with \code{NA}?}
}
\value{
A character vector with leading and trailing whitespace removed.
  If \code{keep_unknown} is \code{FALSE}, blanks and "UNKNOWN" are replaced
  with \code{NA}.
}
\description{
Clean house number and street values
}
