% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-item-cleaning.R
\name{clean_age}
\alias{clean_age}
\title{Clean patient ages}
\usage{
clean_age(age, keep_unknown = FALSE)
}
\arguments{
\item{age}{\code{Age_at_Diagnosis} values.}

\item{keep_unknown}{Replace values for "unknown" with \code{NA}?}
}
\value{
An integer vector of ages.
  If \code{keep_unknown} is \code{FALSE}, values representing unknown ages
  are replaced with \code{NA}.
}
\description{
Clean patient ages
}
