% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-item-cleaning.R
\name{clean_census_block}
\alias{clean_census_block}
\title{Clean Census block group codes}
\usage{
clean_census_block(block, keep_unknown = FALSE)
}
\arguments{
\item{block}{A character vector of Census block group codes.}

\item{keep_unknown}{Replace values for "unknown" with \code{NA}?}
}
\value{
A character vector with leading and trailing whitespace removed.
  If \code{keep_unknown} is \code{FALSE}, blanks and values representing
  unknown block groups are replaced with \code{NA}.
}
\description{
Clean Census block group codes
}
