% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-item-cleaning.R
\name{clean_census_tract}
\alias{clean_census_tract}
\title{Clean Census tract group codes}
\usage{
clean_census_tract(tract, keep_unknown = FALSE)
}
\arguments{
\item{tract}{A character vector of Census tract group codes.}

\item{keep_unknown}{Replace values for "unknown" with \code{NA}?}
}
\value{
A character vector with leading and trailing whitespace removed.
  If \code{keep_unknown} is \code{FALSE}, blanks and values representing
  unknown Census Tracts are replaced with \code{NA}.
}
\description{
Clean Census tract group codes
}
