% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-item-cleaning.R
\name{clean_count}
\alias{clean_count}
\title{Clean counts}
\usage{
clean_count(count, width, keep_unknown = FALSE)
}
\arguments{
\item{count}{A character vector of counts (integer characters only).}

\item{width}{Integer giving the character width of the field.}

\item{keep_unknown}{Replace values for "unknown" with \code{NA}?}
}
\value{
Integer vector of \code{count}.
  If \code{keep_unknown} is \code{FALSE}, values representing unknown counts
  are replaced with \code{NA}.
}
\description{
Replaces any values of all 9's with \code{NA}
(if \code{keep_unknown} is \code{TRUE}) and converts the rest to integers.
}
