% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-item-cleaning.R
\name{clean_county_fips}
\alias{clean_county_fips}
\title{Clean county FIPS codes}
\usage{
clean_county_fips(county, keep_unknown = FALSE)
}
\arguments{
\item{county}{A character vector of county FIPS codes.}

\item{keep_unknown}{Replace values for "unknown" with \code{NA}?}
}
\value{
A character vector with leading and trailing whitespace removed.
  If \code{keep_unknown} is \code{FALSE}, blanks and values representing
  unknown counties are replaced with \code{NA}.
}
\description{
Clean county FIPS codes
}
