% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-item-cleaning.R
\name{clean_facility_id}
\alias{clean_facility_id}
\title{Clean facility identification numbers}
\usage{
clean_facility_id(fin, keep_unknown = FALSE)
}
\arguments{
\item{fin}{A character vector of facility identification numbers (FIN).}

\item{keep_unknown}{Replace values for "unknown" with \code{NA}?}
}
\value{
A character vector with leading and trailing whitespace removed.
  If \code{keep_unknown} is \code{FALSE}, blanks and values representing
  unknown facilities are replaced with \code{NA}.
}
\description{
Clean facility identification numbers
}
