% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-item-cleaning.R
\name{clean_icd_9_cm}
\alias{clean_icd_9_cm}
\title{Clean ICD-9-CM codes}
\usage{
clean_icd_9_cm(code, keep_unknown = FALSE)
}
\arguments{
\item{code}{A character vector of ICD-9-CM codes.}

\item{keep_unknown}{Replace values for "unknown" with \code{NA}?}
}
\value{
A character vector with leading and trailing whitespace removed.
  If \code{keep_unknown} is \code{FALSE}, blanks and the ICD-9-CM code for
  "unknown" (\code{"00000"}) are replaced with \code{NA}.
}
\description{
Clean ICD-9-CM codes
}
