% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-item-cleaning.R
\name{clean_icd_code}
\alias{clean_icd_code}
\title{Clean cause of death codes}
\usage{
clean_icd_code(code, keep_unknown = FALSE)
}
\arguments{
\item{code}{A character vector of ICD-7, ICD-8, ICD-9, and/or ICD-10 codes.}

\item{keep_unknown}{Replace values for "unknown" with \code{NA}?}
}
\value{
A character vector with leading and trailing whitespace removed.
  If \code{keep_unknown} is \code{FALSE}, blanks and the ICD codes for
  "unknown" (\code{"0000"}, \code{"7777"} and \code{"7797"}) are replaced
  with \code{NA}.
}
\description{
Clean cause of death codes
}
