% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-item-cleaning.R
\name{clean_physician_id}
\alias{clean_physician_id}
\title{Clean physician identification numbers}
\usage{
clean_physician_id(physician, keep_unknown = FALSE)
}
\arguments{
\item{physician}{A character vector of medical license number or
facility-generated codes.}

\item{keep_unknown}{Replace values for "unknown" with \code{NA}?}
}
\value{
A character vector with leading and trailing whitespace removed.
  If \code{keep_unknown} is \code{FALSE}, blanks and values representing
  unknown physicians or non-applicable are replaced with \code{NA}.
}
\description{
Clean physician identification numbers
}
