% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-item-cleaning.R
\name{clean_postal}
\alias{clean_postal}
\title{Clean postal codes}
\usage{
clean_postal(postal, keep_unknown = FALSE)
}
\arguments{
\item{postal}{A character vector of postal codes.}

\item{keep_unknown}{Replace values for "unknown" with \code{NA}?}
}
\value{
A character vector with leading and trailing whitespace removed.
  If \code{keep_unknown} is \code{FALSE}, blanks and values representing
  uncertain postal codes are replaced with \code{NA}.
}
\description{
Clean postal codes
}
