% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-item-cleaning.R
\name{clean_ssn}
\alias{clean_ssn}
\title{Clean Social Security ID numbers}
\usage{
clean_ssn(number, keep_unknown = FALSE)
}
\arguments{
\item{number}{A character vector of Social Security identification numbers.
No spaces or punctuation, only numbers.}

\item{keep_unknown}{Replace values for "unknown" with \code{NA}?}
}
\value{
A character vector with leading and trailing whitespace removed.
  If \code{keep_unknown} is \code{FALSE}, blanks and values representing
  unknown Social Security ID numbers are replaced with \code{NA}.
}
\description{
Clean Social Security ID numbers
}
