% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-item-cleaning.R
\name{clean_telephone}
\alias{clean_telephone}
\title{Clean telephone numbers}
\usage{
clean_telephone(number, keep_unknown = FALSE)
}
\arguments{
\item{number}{A character vector of telephone numbers. No spaces or
punctuation, only numbers.}

\item{keep_unknown}{Replace values for "unknown" with \code{NA}?}
}
\value{
A character vector with leading and trailing whitespace removed.
  If \code{keep_unknown} is \code{FALSE}, blanks and values representing
  unknown numbers or patients without a number are replaced with \code{NA}.
}
\description{
Clean telephone numbers
}
