% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-item-cleaning.R
\name{clean_text}
\alias{clean_text}
\title{Clean free-form text}
\usage{
clean_text(text, keep_unknown = FALSE)
}
\arguments{
\item{text}{A character vector of free text values.}

\item{keep_unknown}{Replace values for "unknown" with \code{NA}?}
}
\value{
An character vector with leading and trailing whitespace removed.
  If \code{keep_unknown} is \code{FALSE}, blank values are replaced with
  \code{NA}.
}
\description{
Clean free-form text
}
