% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naaccr_factor.R
\docType{data}
\name{field_levels}
\alias{field_levels}
\alias{field_levels_all}
\title{List of possible values for a field}
\format{A named \code{list}, where the names are for categorical fields or
  sentinel flags, and the values are the possible levels for each field.}
\usage{
field_levels

field_levels_all
}
\description{
These lists gives the levels for each categorical or flag field from the
NAACCR formats. It is intended to help researchers
}
\details{
\code{field_levels} does not include levels representing "unknown."
\code{field_levels_all} does include the "unknown" levels.
}
\keyword{datasets}
