% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/record_format.R
\docType{data}
\name{naaccr_format_12}
\alias{naaccr_format_12}
\alias{naaccr_format_13}
\alias{naaccr_format_14}
\alias{naaccr_format_15}
\alias{naaccr_format_16}
\alias{naaccr_format_18}
\title{Field definitions from all NAACCR format versions}
\format{An object of class \code{record_format} (inherits from \code{data.table}, \code{data.frame}) with 509 rows and 8 columns.}
\usage{
naaccr_format_12

naaccr_format_13

naaccr_format_14

naaccr_format_15

naaccr_format_16

naaccr_format_18
}
\description{
See \code{\link{record_format}}.
}
\details{
A \code{data.table} object defining the fields for each version of NAACCR's
fixed-width record file format.
}
\keyword{datasets}
