% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/record_format.R
\docType{data}
\name{naaccr_formats}
\alias{naaccr_formats}
\alias{naaccr_format_12}
\alias{naaccr_format_13}
\alias{naaccr_format_14}
\alias{naaccr_format_15}
\alias{naaccr_format_16}
\alias{naaccr_format_18}
\alias{naaccr_format_21}
\alias{naaccr_format_22}
\alias{naaccr_format_23}
\title{Field definitions from all NAACCR format versions}
\format{
An object of class \code{list} of length 18.

An object of class \code{record_format} (inherits from \code{data.table}, \code{data.frame}) with 509 rows and 12 columns.

An object of class \code{record_format} (inherits from \code{data.table}, \code{data.frame}) with 529 rows and 12 columns.

An object of class \code{record_format} (inherits from \code{data.table}, \code{data.frame}) with 548 rows and 12 columns.

An object of class \code{record_format} (inherits from \code{data.table}, \code{data.frame}) with 555 rows and 12 columns.

An object of class \code{record_format} (inherits from \code{data.table}, \code{data.frame}) with 587 rows and 12 columns.

An object of class \code{record_format} (inherits from \code{data.table}, \code{data.frame}) with 791 rows and 12 columns.

An object of class \code{record_format} (inherits from \code{data.table}, \code{data.frame}) with 800 rows and 12 columns.

An object of class \code{record_format} (inherits from \code{data.table}, \code{data.frame}) with 810 rows and 12 columns.

An object of class \code{record_format} (inherits from \code{data.table}, \code{data.frame}) with 782 rows and 12 columns.
}
\usage{
naaccr_formats

naaccr_format_12

naaccr_format_13

naaccr_format_14

naaccr_format_15

naaccr_format_16

naaccr_format_18

naaccr_format_21

naaccr_format_22

naaccr_format_23
}
\description{
See \code{\link{record_format}}.
}
\details{
Each \code{naaccr_format_XX} object is a \code{data.table} defining the
fields for each version of NAACCR's record file format.
\code{naaccr_formats} is a list of these record formats, with each name
being the two- or three-digit code for the format.
}
\keyword{datasets}
