% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-item-interpreting.R
\name{naaccr_override}
\alias{naaccr_override}
\title{Interpret basic over-ride flags}
\usage{
naaccr_override(flag)
}
\arguments{
\item{flag}{Character vector of over-ride flags. Its values should only
include \code{""} (blank), \code{"1"}, and possibly \code{NA}.}
}
\value{
A \code{logical} vector with the interpreted values of \code{flag}.
  The interpretation follows these rules: \code{"1"} goes to \code{TRUE}
  (reviewed and confirmed as reported), \code{""} (blank) goes to
  \code{FALSE} (not reviewed or reviewed and corrected), and all other values
  go to \code{NA}.
}
\description{
Interpret basic over-ride flags
}
\examples{
  naaccr_override(c("", "1", NA, "9"))
}
